fix further double conversions

+++ bark_ml/observers/base_observer.hpp
@@ -42,7 +42,7 @@ class BaseObserver {
   virtual ObservedState Observe(
     const ObservedWorld& observed_world) const = 0;
   virtual WorldPtr Reset(const WorldPtr& world) = 0;
-  virtual Box<float> ObservationSpace() const = 0;
+  virtual Box<double> ObservationSpace() const = 0;
   ParamsPtr GetParams() const { return params_; }
 
  private:
